azure-translator-main/requirements.txt
python-dotenv
azure-ai-translation-text==1.0.0b1
flask